# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API with CRUD operations on slugs.",
        "title": "slugs REST API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/slugs": {
            "post": {
                "description": "Create a slug by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Creation of a slug",
                "parameters": [
                    {
                        "description": "Slug object",
                        "name": "slug",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_zpix1_avito-test-task_pkg_entities.Slug"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/slugs/get": {
            "get": {
                "description": "Get slugs by user id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Getting user slugs",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_zpix1_avito-test-task_pkg_entities.GetSlugsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/slugs/history": {
            "get": {
                "description": "Get user slugs history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/csv"
                ],
                "summary": "Get user slugs history in CSV format",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User id",
                        "name": "user_id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Start datetime unixtime (seconds)",
                        "name": "start",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "End datetime unixtime (seconds)",
                        "name": "end",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/slugs/update": {
            "put": {
                "description": "Update user slugs by user id and slug names",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update user slugs",
                "parameters": [
                    {
                        "description": "Slug update object",
                        "name": "slug_update",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_zpix1_avito-test-task_pkg_entities.SlugUpdate"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    }
                }
            }
        },
        "/api/v1/slugs/{slug_name}": {
            "delete": {
                "description": "Delete slug by name",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Deletion of a slug",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Slug name",
                        "name": "slug_name",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/pkg_handler.errorMessage"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_zpix1_avito-test-task_pkg_entities.GetSlugsResponse": {
            "type": "object",
            "properties": {
                "slug_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "github_com_zpix1_avito-test-task_pkg_entities.Slug": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                }
            }
        },
        "github_com_zpix1_avito-test-task_pkg_entities.SlugUpdate": {
            "type": "object",
            "properties": {
                "add_slug_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "delete_slug_names": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "pkg_handler.errorMessage": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}